+++ src/Blitz.php
@@ -107,7 +107,7 @@ class Blitz extends Plugin
     /**
      * @inheritdoc
      */
-    public string $schemaVersion = '4.5.0';
+    public string $schemaVersion = '4.10.0';
 
     /**
      * @inheritdoc